+++ examples/web/src/pages/ClassicApiReferencePage.vue
@@ -1,0 +1,23 @@
+<script setup lang="ts">
+import {
+  ApiReference,
+  type ReferenceConfiguration,
+} from '@scalar/api-reference'
+import { reactive } from 'vue'
+
+const configuration = reactive<ReferenceConfiguration>({
+  theme: 'default',
+  proxy: 'http://localhost:5051',
+  isEditable: false,
+  layout: 'classic',
+  spec: {
+    url: 'https://petstore.swagger.io/v2/swagger.json',
+  },
+  tabs: {
+    initialContent: 'Swagger Editor',
+  },
+})
+</script>
+<template>
+  <ApiReference :configuration="configuration" />
+</template>