+++ packages/api-reference/src/components/Sidebar.vue
@@ -1,21 +1,9 @@
 <script setup lang="ts">
-import { useApiClientStore, useRequestStore } from '@scalar/api-client'
-import { computed, ref, watch } from 'vue'
+import { ref, watch } from 'vue'
 
-import {
-  getApiClientRequest,
-  getHeadingId,
-  getHeadingsFromMarkdown,
-  getLowestHeadingLevel,
-  getModelSectionId,
-  getOperationSectionId,
-  getTagSectionId,
-  hasModels,
-  scrollToId,
-} from '../helpers'
-import { useGlobalStore } from '../stores'
-import { useTemplateStore } from '../stores/template'
-import type { Spec, Tag, TransformedOperation } from '../types'
+import { scrollToId } from '../helpers'
+import { useNavigation } from '../hooks'
+import type { Spec } from '../types'
 import SidebarElement from './SidebarElement.vue'
 import SidebarGroup from './SidebarGroup.vue'
 
@@ -24,167 +12,13 @@ const props = defineProps<{
 }>()
 
 const {
-  state: templateState,
+  items,
+  activeItemId,
   toggleCollapsedSidebarItem,
-  setCollapsedSidebarItem,
-} = useTemplateStore()
-
-// Open the first tag section by default
-watch(
-  props.parsedSpec,
-  () => {
-    const firstTag = props.parsedSpec.tags?.[0]
-
-    if (firstTag) {
-      setCollapsedSidebarItem(getTagSectionId(firstTag), true)
-    }
-  },
-  { immediate: true },
-)
-
-const { server: serverState, authentication: authenticationState } =
-  useGlobalStore()
-
-const { state, toggleApiClient } = useApiClientStore()
-
-const { setActiveRequest } = useRequestStore()
-
-function showItemInClient(operation: TransformedOperation) {
-  const request = getApiClientRequest({
-    serverState: serverState,
-    authenticationState: authenticationState,
-    operation: operation,
-  })
-
-  setActiveRequest(request)
-
-  toggleApiClient(request, true)
-}
-
-const moreThanOneDefaultTag = (tag: Tag) =>
-  props.parsedSpec?.tags?.length !== 1 ||
-  tag?.name !== 'default' ||
-  tag?.description !== ''
-
-const headings = ref<any[]>([])
-
-watch(
-  () => props?.parsedSpec?.info?.description,
-  async () => {
-    const description = props?.parsedSpec?.info?.description
-
-    if (!description) {
-      return []
-    }
-
-    return (headings.value = await updateHeadings(description))
-  },
-)
-
-const updateHeadings = async (description: string) => {
-  const newHeadings = await getHeadingsFromMarkdown(description)
-  const lowestLevel = getLowestHeadingLevel(newHeadings)
-
-  return newHeadings.filter((heading) => heading.depth === lowestLevel)
-}
-
-const isVisible = (id: string) => state.sidebarIdVisibility[id] ?? false
-
-type SidebarEntry = {
-  id: string
-  title: string
-  type: 'Page' | 'Folder'
-  children?: SidebarEntry[]
-  select?: () => void
-  httpVerb?: string
-}
-
-const items = computed((): SidebarEntry[] => {
-  // Introduction
-  const headingEntries: SidebarEntry[] = headings.value.map((heading) => {
-    return {
-      id: getHeadingId(heading),
-      title: heading.value.toUpperCase(),
-      type: 'Page',
-    }
-  })
-
-  // Tags & Operations
-  const firstTag = props?.parsedSpec?.tags?.[0]
-
-  const operationEntries: SidebarEntry[] | undefined =
-    firstTag &&
-    moreThanOneDefaultTag(firstTag) &&
-    firstTag.operations?.length > 0
-      ? props.parsedSpec.tags?.map((tag) => {
-          return {
-            id: getTagSectionId(tag),
-            title: tag.name.toUpperCase(),
-            type: 'Folder',
-            children: tag.operations?.map((operation) => {
-              return {
-                id: getOperationSectionId(operation, tag),
-                title: operation.name,
-                type: 'Page',
-                httpVerb: operation.httpVerb,
-                select: () => {
-                  if (state.showApiClient) {
-                    showItemInClient(operation)
-                  }
-                },
-              }
-            }),
-          }
-        })
-      : firstTag?.operations?.map((operation) => {
-          return {
-            id: getOperationSectionId(operation, firstTag),
-            title: operation.name,
-            type: 'Page',
-            httpVerb: operation.httpVerb,
-            select: () => {
-              if (state.showApiClient) {
-                showItemInClient(operation)
-              }
-            },
-          }
-        })
-
-  // Models
-  const modelEntries: SidebarEntry[] = hasModels(props.parsedSpec)
-    ? [
-        {
-          id: getModelSectionId(),
-          title: 'MODELS',
-          type: 'Folder',
-          children: Object.keys(props.parsedSpec.components?.schemas ?? {}).map(
-            (name) => {
-              return {
-                id: getModelSectionId(name),
-                title: name,
-                type: 'Page',
-              }
-            },
-          ),
-        },
-      ]
-    : []
-
-  return [...headingEntries, ...(operationEntries ?? []), ...modelEntries]
-})
-
-const activeSidebarItemId = computed(() => {
-  const flattenedItems = items.value.reduce((acc, item) => {
-    acc.push(item)
-
-    if (item.children) {
-      acc.push(...item.children)
-    }
-
-    return acc
-  }, [] as SidebarEntry[])
-
-  return flattenedItems.find((item) => isVisible(item.id))?.id ?? null
+  collapsedSidebarItems,
+  sidebarIdVisibility,
+} = useNavigation({
+  parsedSpec: props.parsedSpec,
 })
 
 // This offset determines how far down the sidebar the items scroll
@@ -193,7 +27,7 @@ const scrollerEl = ref<HTMLElement | null>(null)
 const sidebarRefs = ref<{ [key: string]: HTMLElement }>({})
 
 // Watch for the active item changing so we can scroll the sidebar
-watch(activeSidebarItemId, (activeId) => {
+watch(activeItemId, (activeId) => {
   const el = sidebarRefs.value[activeId!]
   if (!el || !scrollerEl.value) return
 
@@ -228,58 +62,64 @@ const setRef = (el: SidebarElementType, id: string) => {
       ref="scrollerEl"
       class="pages custom-scroll custom-scroll-self-contain-overflow">
       <SidebarGroup :level="0">
-        <SidebarElement
+        <template
           v-for="item in items"
-          :key="item.id"
-          :ref="(el) => setRef(el as SidebarElementType, item.id)"
-          data-sidebar-type="heading"
-          :isActive="activeSidebarItemId === item.id"
-          :item="{
-            uid: '',
-            title: item.title,
-            type: item.type,
-            httpVerb: item.httpVerb,
-          }"
-          :open="templateState.collapsedSidebarItems[item.id] ?? false"
-          @select="
-            () => {
-              if (item.id) {
-                scrollToId(item.id)
-              }
+          :key="item.id">
+          <SidebarElement
+            v-if="item.show"
+            :ref="(el) => setRef(el as SidebarElementType, item.id)"
+            data-sidebar-type="heading"
+            :isActive="activeItemId === item.id"
+            :item="{
+              uid: '',
+              title: item.title,
+              type: item.type,
+              httpVerb: item.httpVerb,
+            }"
+            :open="collapsedSidebarItems[item.id] ?? false"
+            @select="
+              () => {
+                if (item.id) {
+                  scrollToId(item.id)
+                }
 
-              if (item.select) {
-                item.select()
+                if (item.select) {
+                  item.select()
+                }
               }
-            }
-          "
-          @toggleOpen="() => toggleCollapsedSidebarItem(item.id)">
-          <template v-if="item.children && item.children?.length > 0">
-            <SidebarGroup :level="0">
-              <SidebarElement
-                v-for="child in item.children"
-                :key="child.id"
-                :ref="(el) => setRef(el as SidebarElementType, child.id)"
-                :isActive="activeSidebarItemId === child.id"
-                :item="{
-                  uid: '',
-                  title: child.title,
-                  type: child.type,
-                  httpVerb: child.httpVerb,
-                }"
-                @select="
-                  () => {
-                    if (child.id) {
-                      scrollToId(child.id)
-                    }
+            "
+            @toggleOpen="() => toggleCollapsedSidebarItem(item.id)">
+            <template v-if="item.children && item.children?.length > 0">
+              <SidebarGroup :level="0">
+                <template
+                  v-for="child in item.children"
+                  :key="child.id">
+                  <SidebarElement
+                    v-if="item.show"
+                    :ref="(el) => setRef(el as SidebarElementType, child.id)"
+                    :isActive="activeItemId === child.id"
+                    :item="{
+                      uid: '',
+                      title: child.title,
+                      type: child.type,
+                      httpVerb: child.httpVerb,
+                    }"
+                    @select="
+                      () => {
+                        if (child.id) {
+                          scrollToId(child.id)
+                        }
 
-                    if (child.select) {
-                      child.select()
-                    }
-                  }
-                " />
-            </SidebarGroup>
-          </template>
-        </SidebarElement>
+                        if (child.select) {
+                          child.select()
+                        }
+                      }
+                    " />
+                </template>
+              </SidebarGroup>
+            </template>
+          </SidebarElement>
+        </template>
       </SidebarGroup>
     </div>
     <slot name="sidebar-end" />