+++ packages/api-reference/vite.standalone.config.ts
@@ -33,7 +33,7 @@ export default defineConfig({
       include: [/@scalar\/swagger-editor/, /node_modules/],
     },
     cssCodeSplit: false,
-    minify: false,
+    minify: 'terser',
     lib: {
       entry: ['src/standalone.ts'],
       name: '@scalar/api-reference',