+++ packages/fastify-api-reference/vite.config.ts
@@ -1,4 +1,5 @@
 import path from 'path'
+import { viteStaticCopy } from 'vite-plugin-static-copy'
 import { defineConfig } from 'vitest/config'
 
 import pkg from './package.json'
@@ -6,7 +7,18 @@ import { nodeExternals } from './src/vite-plugins'
 import { nodeShims } from './src/vite-plugins'
 
 export default defineConfig({
-  plugins: [nodeShims(), nodeExternals()],
+  plugins: [
+    nodeShims(),
+    nodeExternals(),
+    viteStaticCopy({
+      targets: [
+        {
+          src: '../api-reference/dist/browser/standalone.js',
+          dest: './js',
+        },
+      ],
+    }),
+  ],
   build: {
     // If minify is enabled, the nodeShims extension doesn’t work.
     minify: 'terser',